Fail gracefully when we can't do a native build.
Before this CL, attempting to build (for example) a Windows wheel on
Linux would crash the process with 'Cannot identify native image for
linux2'. Avoid that crash by returning an empty native platform list on
Linux, and also handle the case where the platform we're attempting to
build isn't in the list.

With this change we'll just move on to the next wheel rather than
crashing the process.

Bug: 1144965
Change-Id: Ibb9463fd0f6ae2526d0ee47553e8471cbcc4b48e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2520359
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#36199}