chromeperf: Implement a token deletion pipeline
This change introduces the upload deletion token Dataflow pipeline which
is aware of the relationship between `Token` and `Measurement` entities
in Datastore. This pipeline is required to manage the amount of entities
related to the processed uploads, which can grow unbounded as we ingest
more data day-by-day.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1128357
Change-Id: I5737af9fede4c4a1a85af0386d7bb01f59fdec35
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2520363
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36226}