Add a debug sleep at the entry point of dockerbuild.
The Windows builder fails in a hard-to-debug manner, and I don't have
the permissions to log into the bots or to use led to launch builds with
locally-modified code, so unfortunately it seems this is the easiest way
to debug it.

Right now the step that runs dockerbuild "succeeds" after 500ms with no
output. It looks like it fails to close the stdin/stdout fds for some
reason. This sleep is an experiment to see if we can get any output from
it at all.

Bug: 1144965
Change-Id: Iec96d2d8152ec9e82ab95f1f99b326c13027ae1d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2525381
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#36248}