[3pp] fix pip_bootstrap with supplying artifact names.
For packages that uses pip_bootstrap, it performs a wheel name check,
whereas we name downloaded source artifacts in a generic way.

This is a single package where name of the artifact is important and is
used as part of installation. We need to introduce passing artifacts
name via download manifest.

In this CL, we add that and make it optional since this is not commonly
used across other packages.

Bug: 1131685
Change-Id: I8de0c4cf89db91e4c8b75e18554d382a33ea78dc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2538644
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Auto-Submit: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#36439}