[provenance] Added asymmetric sign for generator.
Provenance generator will use provenancepb to serialize data between
builder and invoker of the binary. This will be signed with key stored
in cloudkms.

Although, there is a cli command in luci that supports signing ops we
use today, to remove any dependency on luci, here's a quick
implementation of signing operation with cloud kms as documented on
https://cloud.google.com/kms/docs/create-validate-signatures#kms-sign-asymmetric-go.

Due to the return type of cloud kms signature is DER formatted but jwt
uses a different/raw format, we might need to do some tweaking to the
signature before using it.

R=jclinton@google.com

Bug: 1093088
Change-Id: Iaa7005470bc34199052f80e9976c03b93ee4fa9d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2542741
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36552}