[weblayer, skew tests] Refactor recipe so that it will refresh skew tests nightly
This recipe will be triggered by the chrome-release builder. It will
receive a chromium version as an input property. If the version is a
tag in one of last 4 milestone branches then it will refresh skew tests
for that branch using the input version. This recipe copies code from
build_weblayer_version_tests_apk_cipd_pkg.py. Eventually that recipe
will be deprecated and this recipe will replace it.

Bug: 1147459, 1099017, 1041619
Change-Id: I1560308fd307ccc113178272580099e398f19f92
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2548843
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Michael Moss <mmoss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36585}