Only check lines after encountering the histogram start pattern
Only check lines after encountering the histogram start pattern. This is
mainly for preventing new variants's owners tag from causing Tricium to
crash. But it also improves the efficient of the Tricium as it won't try
to do complex string matching before entering a <histogram> tag.

Another case that can cause the Tricium to fail is when the <histogram>
tag is multi-lined. That case will be addressed in a follow-up cl.

Bug: 1147651
Change-Id: I5821e2c122689b2506c56d59dce7440cd242dc8f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2545298
Commit-Queue: Weilun Shi <sweilun@chromium.org>
Reviewed-by: Quinten Yearsley <qyearsley@chromium.org>
Auto-Submit: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36607}