chromeperf: Use RawURLEncoding.DecodeString
This change uses the more permissive (non-padding requiring) base64
encoder/decoder for the payload of the authorization header that's
stored in the x-endpoint-api-userinfo header.

We also switch to using an opaque map for JSON parsing so that we can
get all the fields (not just the 'email' field) for later authorization
checks.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1059667
Change-Id: Ib5bce574e2fd722120d4b70842df1b04e4f454d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2560420
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Cr-Commit-Position: refs/heads/master@{#36722}