chromeperf: Handle responses from legacy service
This change makes the server handle the HTTP result codes from the
legacy service, and propagates the issue to the client. This code is
only useful in the interim because we're proxying requests to the legacy
service. In the future these error messages will not make it to the
client, as we'll have a better logging and debugging strategy in
production for this gRPC service.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1059667
Change-Id: I2b245420687538bdc3418c7244dba1224603ed99
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2560425
Commit-Queue: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Cr-Commit-Position: refs/heads/master@{#36735}