[3pp] Move git tag movement to error from step failure.
Resolved todos left on 3pp recipe that were blocked my migration to new
fetch raw workflow.

One of the todo was to move to raising error and stop recipe execution
if a git tag movement is detected. Earlier, recipe would raise a step
failure and contune with build. But since git tag movement might be an
indicator of compromise for a remote git repository we are relying on,
it would be better to raise an error to halt recipe execution.

This should be pretty rare and should not impact current flow. Tag
movement of a git repository is very suspicious and almost should never
happen.

R=bryner@google.com, jclinton@google.com

Bug: 1131685
Change-Id: I96fb3443ebe6b7b49c31c103b65c891cae670d5a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2561754
Commit-Queue: Brian Ryner <bryner@google.com>
Auto-Submit: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#36743}