chromeperf: Extract selection from deletion
In this change we extract the selection phase of the token deletion
pipeline to allow us to test our assumptions on the processing we're
doing in the pipeline. We add some light testing of the pipeline of our
assumptions with some in-memory data.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1155096
Change-Id: I5f3c644b86e13c3ae5b6de232c2edc84970fa9bb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2570757
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36893}