[provenance] Fixed a minor issue with formatting output attestation.
To build a json and write it to a file, we need to declare the field
names in the Go struct definition. Earlier we did not have that and the
cli was generating empty attestation.

We have added a field `Jwt` and tagged it as "jwt" in the created json
byte. This is tested and works as expected.

R=jclinton@google.com

Bug: 1093088
Change-Id: I5db0a68ec0529a929471deed88edde19692344d5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2577955
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Commit-Queue: Akash Mukherjee <akashmukherjee@google.com>
Cr-Commit-Position: refs/heads/master@{#36958}