[provenance] Publish provenance package to cipd.
Although provenance binaries will be embedded into bot OS images,
pushing it to CIPD is still a good idea to aid developers into using and
testing with this tool to attach and verify (in future) provenance
metadata to built artifacts.

This will provide developers a CLI to play with attestation and
understand what is required to on-board with BCID Verified Builds for
Software initiative.

R=jclinton@google.com

Bug: 1093088
Change-Id: I44fd27198c4344b5024c6d6065af2fed5c23710c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2582490
Auto-Submit: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#36999}