phosphorus: Write out crashes and compare with RTD
Write crashes generated by FetchCrashes to test_runner/ output directory
(for an eventual replacement of the RTDs' logic to do this) and add
logic to compare crashes written by the RTDs to those written by
FetchCrashes.

In order to verify that FetchCrashes is reliable, add output comparing
it against the crashes found by the RTD.

To test, I manually created some chrome crashes in one of the RTD
folders, some crashes on the DUT (debugd and anomaly_detector), put the debugd
crash in the RTD folder, and ran phosphorus fetch-crashes, with
the below output.

{
   "crashes" : [
      {
         "execName" : "anomaly_detector",
         "filenameBase" : "anomaly_detector.20201207.180210.85651.2518"
      },
      {
         "execName" : "debugd",
         "filenameBase" : "debugd.20201208.154043.91203.1516"
      }
   ],
   "crashesRtdOnly" : [
      "chrome.20201203.165649.34580.21118",
      "chrome.20201203.165649.48675.21122",
      "chrome.20201203.165649.60782.21116",
      "chrome.20201203.165649.69032.21074",
      "chrome.20201207.120207.67088.17754"
   ],
   "crashesTlsOnly" : [
      "anomaly_detector.20201207.180210.85651.2518"
   ],
   "state" : "SUCCEEDED"
}

BUG=chromium:1107005
TEST=run fetch-crashes; check output (see above)

Change-Id: I23cec6ff935c9e59583ec3f472142eb151dcd9ea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2579670
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37037}