phosphorus: Add build-parallels-image-save command.
Provide a means for build_parallels_image to directly update the DUT
state in UFS, so that it can mark DUTs as ready or needing repair.
Autotest local state is not used or maintained in this recipe, so
using the skylab_local_state command seemed inappropriate.

Uses protos proposed in crrev.com/c/2580921

BUG=chromium:1105967
TEST=build/build.py

Change-Id: Ic9d3577adb31ada00e009cd0a09e5ee65b05460d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2580662
Commit-Queue: Patrick Meiring <meiring@google.com>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37069}