[cr-audit-commits] Nag the author of TBR'd CLs with the auto-filed bug
We've been nagging via posted messages to the TBR CLs 1 day after a TBR
is committed for awhile now. It's time to make those nags immediate and
more critical.

Since escalting inside of the CL is politically difficult, we'll instead
assign the bug to the author of the CL and then, if no action is taken,
escalate further.

This CL also completely deletes the concept of a reminder post to
Gerrit.

Bug: chromium:1092606
Change-Id: I29c794ef6596aa33e3b1012052b1c43e8128c883
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2574096
Commit-Queue: Jason Clinton <jclinton@chromium.org>
Reviewed-by: Joanna Wang <jojwang@chromium.org>
Reviewed-by: Xinyu Yang <xinyuoffline@google.com>
Reviewed-by: Yulan Lin <yulanlin@google.com>
Auto-Submit: Jason Clinton <jclinton@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37101}