Reference build/ lkgr_finder recipe for chromium-lkgr-finder
In support of inverting the infra/infra -> build dependency,
the lkgr_finder has been moved to the build/ repository.
This CL updates the builder to reference the build/ recipe such
that we can deprecate the one here, which removes the
dep to chromium_checkout.

Bug: 1157953
Change-Id: I268fb264dd747c40ccdee26a890e4318781efd2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2593364
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Cr-Commit-Position: refs/heads/master@{#37119}