tricium: Make `make` build analyzers for linux by default
The idea is, if anyone modifies any analyzers in the
on a Mac or other platform, they can run `make test`
and `go build` and the built executable will run;
but if they run ./update.sh to update the version in
CIPD, then a linux binary should be uploaded.

Note: This also updates the spellchecker dictionary,
which is a trivial and safe change.

Bug: 1033579
Change-Id: I8bea3453a6cb84b1064d8334884f63e3616c770c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2590164
Auto-Submit: Quinten Yearsley <qyearsley@chromium.org>
Reviewed-by: Marc-Antoine Ruel <maruel@chromium.org>
Commit-Queue: Quinten Yearsley <qyearsley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37172}