phosphorus: Add logging to fetch-crashes command.
https://dashboards.corp.google.com/_dbfdd634_b093_49b5_9c65_49f85cf4b4b0
shows rare cases where fetch-crashes runs for an exceedingly long time.

https://luci-milo.appspot.com/ui/p/chromeos/builders/test_runner/test_runner/b8861152505129815392/steps?succeeded=true&debug=false
is an example of a long execution time --- fetch-crashes indicates that
it's running with a ~16-hour timeout:

[I2020-12-12T03:54:14.940890Z 604554 0 fetch-crashes.go:92] Running with deadline 2020-12-12 19:54:14 +0000 UTC (current time: 2020-12-12 03:54:14.94070739 +0000 UTC)

In this case, fetch-crashes ended up taking 13 hours 47 minutes to
complete, so it did not hit that deadline.

This logging should help us figure out where fetch-crashes is getting
stuck.

In a separate CL, we'll specify a shorter deadline for it.

BUG=chromium:1107005
TEST=ran phosphorus fetch-crashes, verified that logs appeared.

Change-Id: Ia8c50c16cee30ec308f8e47a0419270c00f65a31
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2597820
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Reviewed-by: Prathmesh Prabhu <pprabhu@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37185}