[Findit][Isolate] Return test result object instead of dict in swarmed_test_util.GetSwarmingTaskDataAndResult
swarmed_test_util.GetSwarmingTaskDataAndResult() currently returns a dict for test results, which makes it harder to convert to resultDB, since resultDB does not have a dict.

It would make it easier if it returns a TestResults object, so later we can plug in the ResultDBTestResults.

Bug: 981066
Change-Id: I3237c03ee769eab3d3e410cc7071e0bd2de9135f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2617058
Commit-Queue: Quang Minh Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Joey Scarr <jsca@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37448}