Fix a bug with packages that use a host bootstrap version.
Some packages depend on a host-built version of the same package when
cross-compiling for another architecture. In this case, there will
be a ResolvedSpec loaded for both of the platforms. De-dup by package
name before copying the package definitions into the checkout, to avoid
errors trying to overwrite existing files.

Bug: 1152854
Change-Id: Id2cc4831e2b5b7efae263bd6dfa1b33bc7a12f6d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2635962
Reviewed-by: Owen Rodley <orodley@chromium.org>
Reviewed-by: Haiyang Pan <hypan@google.com>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#37763}