Update deployment documentation for objective-c-style
This updates the README.md file to replace the reference to
"go build" in the Deployment section with "make", and to
suggest using "update.sh" instead.

"make" sets the target OS and architecture to match the bots,
while "go build" builds for the local platform, which might
not necessarily match the bots.

Bug: None
Change-Id: I435d4db6e6e8cfac10fdb28288afb03a39daa810
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2643223
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#37854}