[UFS] Fix race condition in SyncMachinesFromAssets
Test: make test

Fixes the race condition during SyncMachinesFromAssets job. As the job
processes all the assets in the table it's not practical to run it as
a transaction. This means deleting an asset at the right time can cause
machine to be created even though the asset is deleted. Checking for
asset existence before creating machine seems to fix this issue because
AssetRegistration is run as a transaction.

Change-Id: Idb8deb8a74a004ce849f7ac68b4e971bfbe03f49
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2664950
Reviewed-by: Eshwar Narayan <eshwarn@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38123}