phosphorus: Enable SSP for all client tests
We should've always done this.

When Tauto executes client tests:
* control file is read from the drone
* Python test content is read from the DUT

If Server Side Packaging is disabled, the control file read from the
drone is not the same version as the target Chrome OS, and hence the
python test content.

This is anyway a step in the right direction: We want *all* tests to
always run containerized.

Bug: 1109600
Change-Id: I457d16fa0f472fc12a69877c32db015de19a5050
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2373369
Reviewed-by: Congbin Guo <guocb@chromium.org>
Reviewed-by: Sean Abraham <seanabraham@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@google.com>
Cr-Commit-Position: refs/heads/master@{#38329}