Manual Repair: implement host status check for batch process ui
When entering a comma-separated list of hostnames, the UI will check if
the host exists in Inv v2 and also if the host has an open record. If
the device does not exist of if the host has an open record, then the
user will not be able to create a record for that host.

Screenshot: https://screenshot.googleplex.com/5skyxtdSVbfug8h.png

BUG=chromium:1145259
TEST=npm run dev

Change-Id: I7350317da39b6fa4dd6e6167685007bc7fda7c33
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2680164
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38434}