Manual Repair: implement repair form for batch creation
A repair form similar to the single record creation form is implemented
for batch creation. We can slim down the form if necessary to minimize
overhead for repair team. UI allows all inputted host to have a record
created based on input from repair form. Once submitted, a results
dashboard is shown with record creation status and relevant error
messages if any.

Screenshots:
- Repair form: https://screenshot.googleplex.com/7Dr3wddhY7E5cFR.png
- Results panel: https://screenshot.googleplex.com/8kJNQGFshtecZy9.png

Note that the form code is not currently refactored as we do not know
the direction of the project so it is best to keep the forms decoupled
for now. Components can be revised and shared once we determine UI
needs.

BUG=chromium:1145259
TEST=npm run dev; functionality tested in browser

Change-Id: I405fc40bf8646828e89fad689624405aa16d5b40
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2682059
Commit-Queue: Justin Suen <justinsuen@google.com>
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38438}