[rts] Define a range for the number of changed files
https://chromium-review.googlesource.com/c/chromium/src/+/2644429/17
modified 7K+ files, but
https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/618684/overview
reported none. This is probably a bug in the recipe implementation,
but the pipeline is better be resilient to such bugs by ignoring builds
that don't report changed files.

Similarly, it is more risky to enable RTS if a lot of files have been
changed, and less likely that the selection strategy can prune a lot of
tests.

Define a range for the number of changed files, as [1, 30].
Note that 96.8% of commits changed <30 files.

Use this range when fetching evaluation data, and when selecting tests.
Note that its OK not to include constants into the model directly,
because the binary itself is added to the model, so constants are included
indirectly.

Also make casing of BQ parameter names more consistent, by using only
camelCase.

Bug: 1090668
Change-Id: Ia794dbc200f91e8e607e0114f62ce50dae61776e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2704969
Reviewed-by: Gregory Guterman <guterman@google.com>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38639}