pinpoint: Expand platform list
The previous change had too short a list of platforms to build and
deploy the Pinpoint CLI. This change ensures that we're building in as
many platforms as we're supporting.

This copies the list of platforms from the `bb.yaml` config.

R=chowski@google.com, nodir@chromium.org

Bug: chromium:1175615
Change-Id: I79c42a902c0ccb2eee14380b50578a08713d90f7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2712122
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38701}