Add an 'env' field on Platform.
This allows platforms to have extra environment variables that are set
for building wheels.

Now that we have this, we can move some of the environment variables
that we set in the wheel building recipe here instead. That way they'll
be set when we run dockerbuild locally too.

We don't want to move the Windows distutils variables though, as those
are specific to how we set up MSVC on the builders.

Bug: 1181053
Change-Id: I148a861a93b105baebf1d31a9ad8fe61e835d37e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2712315
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Brian Ryner <bryner@google.com>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38726}