[go] Replace INFRA_GO_VERSION with INFRA_GO_VERSION_VARIANT.
We want to pin specific Go versions we support across all
environments in the single bootstrap.py file. So instead of
telling it what concrete version to install, we tell it
that we want just a specific variant, with options "default",
"legacy" and "bleeding_edge".

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1102967

Change-Id: Ibd7307d36a16c717fb9b017732020bfd580e4625
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2713737
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38741}