pinpoint cli: create a wrapper for our subcommands.
In idiomatic Go code, any function which needs a context or has an error
to deal with should, generally, take a context as the first parameter
and return an error as its last return parameter. This CL introduces a
wrapper type which allows us to do so while still using the subcommands
framework. In effect:

* cli.GetContext is only called in one place
* Commands don't need to take 'env' as a parameter
* Eliminates the baseCommandRun.done helper method

This CL also converges how flags are handled in the various base
commands and command implementations: via a RegisterFlags method.

Test: ran wait-job, get-job, list-job
Change-Id: I32b90d72a0b05675515a33aad8dde21b68cd81ee
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2714523
Reviewed-by: Dean Berris <dberris@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Kevin Chowski <chowski@google.com>
Cr-Commit-Position: refs/heads/master@{#38750}