[result_adapter] Use text-artifact
Embed the artifact instead of asking the user to open it.

This does not solve the problem fully: this "stack_trace" artifact is actually
HTML, so when embedding it is printed as HTML. To fix it properly, we still need
crbug.com/1149662 to be fixed.

Bug: 1171620
Change-Id: I6a3bc0b2be1bfa6a48c5c18ba38cd60ef694c656
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2706268
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38789}