[cfggrab] Add a primitive tool to help grepping LUCI configs.
It is often useful to quickly grep through all project configs of
some particular LUCI service, e.g. when refactoring config validation.

`migrator` tool can be used for that, but it is very heavy and
requires the caller to have git-level access to all configs (which LUCI
developers don't have).

cfggrab allows to quickly do stuff like:

  $ cfggrab luci-scheduler.cfg | grep "bucket: "

To see what formats are used for the "bucket" field in the wild.

R=tandrii@chromium.org, iannucci@chromium.org

Change-Id: Icff892ca8b14305ceed830486ae288f9dacb55d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2718724
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38800}