phosphorus: allow more GS retries before failing permanently
Currently we're seeing hundreds-to-thousands of test_runners dying each
day because they run out of retries in uploading files to GS. The
problem is that we're uploading 20 files at once, so if there's a
momentary networking break, then all 20 of those uploads will fail and
we'll immediately run out of retries, since we only permitted 10 retries
across all uploads. Bumping the number to 100 will allow all 20 of those
to fail and then for all of those to actually be retried.

BUG=chromium:1168728
TEST=unit tests

Change-Id: If69095e258ae94adf6e9ccfc689885b41617497e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2727777
Auto-Submit: Sean Abraham <seanabraham@chromium.org>
Reviewed-by: Taylor Clark <taylormade@google.com>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38902}