pinpoint: support downloading results via get-job and wait-job.
Downloading results is off by default, and is enabled via the
-download-results flag. There is support for specifying which directory
to store results to, as well as a convenience flag for opening up the
results with the default application (using xdg-open). More ways to do
that (e.g. for other OSs) could be supported in the future as needed.

This CL adds a downloadResultsMixin, which is like a sub-subcommand to
encapsulate functionality that is common among different commands. A
"mixin"-like thing felt like a better match than doing an
"inheritence"-like thing because downloading results might be desired in
any sort of command, including ones which start experiments.

TEST: Ran get-job with already-finished job '119177a9d20000'
Change-Id: I1f54138db7fbfa0c42caa60ec551d29c13f7f96e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2717545
Reviewed-by: Dean Berris <dberris@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Kevin Chowski <chowski@google.com>
Cr-Commit-Position: refs/heads/master@{#38955}