support_3pp: Introduce a new Manifest type "script".
This is similar to the old flow for ScriptSource where "checkout" is
called to download the artifacts.
This will be enabled only when "use_fetch_checkout_workflow" is set
in the 3pp.pb

The reason to add this is that some artifacts may not have explicit URL
for download, but instead may provide APIs for downloading. But the
current "get_download_url" flow does not support this.

Bug: 1185395
Change-Id: I4f2e9b6a78b29ee370372493042e73c993888804
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2713533
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#39214}