Improve the "get_url" for ScriptSource in support_3pp.
In ScriptSource, the "ext" is always required, even it is not used
at all when "name" is specified.

This CL improves the logic so that it makes both "ext" and "name"
optional and errors out when both are empty.

Also updates the step_test_data to make it more clear about the
expected format for each key.

Change-Id: I05eb8a14572790dba500101a526ade29dba4d512
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2745788
Commit-Queue: Haiyang Pan <hypan@google.com>
Reviewed-by: Akash Mukherjee <akashmukherjee@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#39216}