Pinpoint: first stab at a fake Legacy Pinpoint implementation.
Just the bare-bones of functionality: you can preload job data, and call
the `/api/job/$ID` API. JSON data is served by applying the in-memory
representation of a Job to a text template.

The single example JSON template (job.tmpl) was copied from the testdata
of the gRPC server. If needed, the process of gathering example data for
faking out can be automated more in the future.

To test this, instead of duplicating the Legacy API constraints again, I
thought we might be able to get away with reusing the new Pinpoint
service as a controller; definitely channeling "prefer real objects"
here. Advice on better/additional ways to test this would be welcome.

This CL also moved some flags defined at package-scope into the Main
function, to avoid flag pollution in the fakelegacy_test which imports
that package.


Change-Id: I50c1e0ea4f2e2a2cfbf8ee286715f01918a88b10
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2769286
Commit-Queue: Kevin Chowski <chowski@google.com>
Reviewed-by: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#39533}