tls: support block-level wiping stateful partition
Adds support in TLS for wiping the stateful partition.
If ProvisionDutRequest.preserve_stateful is not set, the provisioning
code does an extra step to block-level wipe and reset the stateful
partition based on the newly provisioned image. This is done using
a clobber-state invocation that adds another reboot to the process.

BUG=b:182822167
TEST=Test provisioning with TLS+TLW; check stateful partition is
     wiped and recreated (dumpe2fs).

Cq-Depend: chromium:2762544
Change-Id: I4644bcbb68fbc49e1224a73fe81f73539258fa98
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2762517
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#39631}