Pinpoint cli: infer some flag defaults from `git cl issue`.
Added in https://crrev.com/c/2766153, we can use the structured output
of that command to guess which gerrit/gitiles host to use, as well as
the repository the user cares about.

If an error occurs, we silently fall back to the hard-coded default of
Chromium.

Change-Id: Icce6a508585b10d0893a19d30f06b6382c1d633d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2783360
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>
Commit-Queue: Kevin Chowski <chowski@google.com>
Cr-Commit-Position: refs/heads/master@{#39693}