3PP: Don't upload if there are multiple instances with the tag.
Currently we take any failure of "cipd describe" to mean that the
package doesn't exist. But "cipd describe" can also fail in the case
where the reference is ambiguous due to multiple instances having the
tag.

This can come about due to a race between different 3PP builders
building and uploading the same package. If it happens we'll get stuck
in a loop of uploading more and more packages with the same tag.

Here we fix that by explicitly checking for this condition, and not
uploading the package again in that case.

Bug: 1186895
Change-Id: I4389d4aafd43321b114e8938c1f8c961cb748ce7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2790205
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#39757}