phosphorus: Use Timestamp.IsValid() instead of .AsTime().IsZero()
It looks like a recent change in proto behaviour has caused
Timestamp.AsTime() to return the unix epoch (1 January 1970)
if Timestamp is nil, instead of golang's zero time (00:00:00 UTC
on 1 January, year 1). This appears to be 'correct' with respect
to the Timestamp proto (which encodes seconds and nanoseconds
since the unix epoch) but has broken code dependent on the old
behaviour.

BUG=chromium:1193349
Test=None

Change-Id: Ibfae38a2a32dcb9afb5475029fa7fa0009e0fd12
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2790195
Reviewed-by: Dhanya Ganesh <dhanyaganesh@chromium.org>
Reviewed-by: Jack Neus <jackneus@google.com>
Commit-Queue: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/master@{#39826}