Pinpoint: introduce fakelegacy-based smoke test.
The test relies on using a local infra dev environment to compile the
gRPC server, CLI, and fakelegacy service. The fakelegacy and gRPC
service are started, with the latter pointing at the former, and each
flow executes the CLI and inspects the output.

The fakelegacy server only supports getting, listing, and starting jobs,
so those flows are tested here.

Output inspection on the CLI is a bit rough since it's human-readable,
mostly unstructured output. In the future, if the CLI gains a -json flag
or some other more structured output format, this code can be revisited
for the better.

Change-Id: I8bf8b13b24d244855f065b5e6613e5058f0b222a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2794926
Commit-Queue: Kevin Chowski <chowski@google.com>
Reviewed-by: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#39845}