[buildbucket/py] Add luci.swarming.use_rbe_cas temporarially.
This will allow us to roll out the new experiment support CLs today; We
can rename the experiment in parallel.

In particular, this does NOT ensure that "luci.swarming.use_rbe_cas"
is marked as negative on builds which don't have it, since we're
planning to rename it.

R=smut, tikuta, vadimsh, yiwzhang, yuanjunh

Bug: 1093970,1143122
Change-Id: I84c3b9e3fb292ff88e084c6e258dc89600600abb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2807534
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#39946}