caching-backend: allow customization on interface keepalived listens on
We don't need this feature in standard k8s since all nodes are
configured with bond0. This feature is only useful when bond0 is missing
and we cannot wait for a day for lab team to create it.

BUG=None
TEST=Ran locally to ensure the conf generated looks good.

Change-Id: Ie675c3eb95c895b9e1411b2dea474221ffdb7fc2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2808705
Commit-Queue: Congbin Guo <guocb@chromium.org>
Auto-Submit: Congbin Guo <guocb@chromium.org>
Reviewed-by: Sanika Kulkarni <sanikak@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/master@{#39959}