fetch-crashes: Don't examine provision directory.
fetch-crashes attempts to locate all crashes that the RTD found but that
the TLS API missed. One directory it examines is
`provision_*/crashinfo.*/`. However, we expect that the provision step
will clear crashes before the test runs and we don't need the
FetchCrashes TLS API to upload these crashes. So, stop checking it.

BUG=chromium:1107005
TEST=CQ

Change-Id: I36ef306dd56b8328501552ccf83dde26f7edfea7
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2815898
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40057}