Compute lists of build targets based on a SourceTestPlan.
- Currently, the output is just a Go struct, that has a name
(e.g. "kernel-4.14"), list of BuildTargets, and test tag
information. It is my understanding that the input to CTP v2
will contain similar information, using this as a placeholder
until the interface for CTP v2 is formally defined. At that
time, this function will be changed.

- Gets kernel, SoC, and build target information from a
BuildSummaryList.

BUG=b:182898188
TEST=CQ

Change-Id: Ic6da5359bfb03854c747e24db622f50818779c07
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2822049
Reviewed-by: jaquesc <jaquesc@google.com>
Reviewed-by: Sean McAllister <smcallis@google.com>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40086}