[dirmd] Add Metadata.imports field
The Metadata.inherit_from field design turned out to be flawed.
Mostly because it applies to *all* attributes, as opposed only to the
attributes that the developer intends to reuse. As a result, usage of
inherit_from causes an unintended loss of inherited metadata, e.g.
inheriting android-specific code from non-android specific causes
the inheretee to lose `os: Android` metadata defined in the ancestor.

Introduce Metadata.imports which does not affect inheritance at all.
It is more directly aligned with the problem we are trying to solve,
which is: certain attribute values are common and ideally are defined
in one place. Then other places can simply import from there.

Bug: 1179786
Change-Id: I248acb16573b4f7310e4ef7cebd830a76c972f8d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2832491
Reviewed-by: Josip Sokcevic <sokcevic@google.com>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40357}